feat: replace next-sitemap config with ESM
